% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_matrix.R
\name{a_p}
\alias{a_p}
\title{Function a_p.}
\usage{
a_p(p)
}
\arguments{
\item{p}{power, real number from (-1,1)}
}
\description{
Computes the corresponding function value from Mazur et al. 2018.
}
\references{
\insertRef{MOP18}{rlfsm}
}
