% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_estimates.r
\name{H_hat}
\alias{H_hat}
\title{Statistical estimator of H in high/low frequency setting}
\usage{
H_hat(p, k, path)
}
\arguments{
\item{p}{power}

\item{k}{increment order}

\item{path}{sample path of lfsm on which the inference is to be performed}
}
\description{
The statistic is defined as
\deqn{\widehat{H}_{\textnormal{high}} (p,k)_n:= \frac 1p \log_2  R_{\textnormal{high}} (p,k)_n,
\qquad \widehat{H}_{\textnormal{low}} (p,k)_n:= \frac 1p \log_2  R_{\textnormal{low}} (p,k)_n}
}
\references{
\insertRef{MOP18}{rlfsm}
}
