% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_estimates.r
\name{m_pk}
\alias{m_pk}
\title{m(-p,k)}
\usage{
m_pk(k, p, alpha, H, sigma)
}
\arguments{
\item{k}{increment order}

\item{p}{a positive number}

\item{alpha}{self-similarity parameter of alpha stable random motion.}

\item{H}{Hurst parameter}

\item{sigma}{Scale parameter of lfsm}
}
\description{
defined as \eqn{m_{p,k} := E[|\Delta_{k,k} X|^p]} for positive powers.  When p is negative (-p is positive) the equality does not hold.
}
\details{
The following identity is used for computations: \deqn{m_{-p,k} =  \frac{(\sigma \|h_k\|_{\alpha})^{-p}}{a_{-p}} \int_{\R} \exp(-|y|^{\alpha}) |y|^{-1+p} dy = \frac{2(\sigma \|h_k\|_{\alpha})^{-p}}{\alpha a_{-p}} \Gamma(p/\alpha)}
}
\references{
\insertRef{MOP18}{rlfsm}
}
