% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_h_kr.R
\name{Norm_alpha}
\alias{Norm_alpha}
\title{Alpha-norm of an arbitrary function}
\usage{
Norm_alpha(fun, alpha, ...)
}
\arguments{
\item{fun}{a function to compute a norm}

\item{alpha}{self-similarity parameter of alpha stable random motion.}

\item{...}{a set of parameters to pass to integrate}
}
\description{
Alpha-norm of an arbitrary function
}
\details{
fun must accept a vector of values for evaluation. See ?integrate for further details.
Most problems with this function appear because of rather high precision. Try to tune rel.tol parameter first.
}
\examples{
Norm_alpha(h_kr,alpha=1.8,k=2,r=1,H=0.8,l=4)
}
\references{
\insertRef{MOP18}{rlfsm}
}
