% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_theorems.R
\name{phi_of_alpha}
\alias{phi_of_alpha}
\title{Inverse alpha estimator}
\usage{
phi_of_alpha(p, p_prime, alpha)
}
\arguments{
\item{p}{power}

\item{p_prime}{power}

\item{alpha}{self-similarity parameter of alpha stable random motion.}
}
\description{
A function from a general estimation procedure which is defined as m^p_-p'_k /m^p'_-p_k, originally proposed in [13].
}
\references{
\insertRef{MOP18}{rlfsm}
}
