% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLotting_sample_paths_stable.R
\name{Plot_list_paths}
\alias{Plot_list_paths}
\title{Rendering of path lattice}
\usage{
Plot_list_paths(arr)
}
\arguments{
\item{arr}{a data frame produced by \code{\link{Path_array}}.}
}
\description{
Rendering of path lattice
}
\examples{
l=list(H=c(0.2,0.8),alpha=c(1,1.8), freq="H")
arr<-Path_array(N=300,m=30,M=100,l=l,sigma=0.3)
p<-Plot_list_paths(arr)
p
}
