% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_matrix.R
\name{U_g}
\alias{U_g}
\title{alpha norm of u*g}
\usage{
U_g(g, u, ...)
}
\arguments{
\item{g}{functions \eqn{g,h: \R to \R} with finite alpha-norm (see \code{\link{Norm_alpha}}).}

\item{u}{real numbers}

\item{...}{additional parameters to pass to Norm_alpha}
}
\description{
alpha norm of u*g
}
\examples{
g<-function(x) exp(-x^2)
g<-function(x) exp(-abs(x))
U_g(g=g,u=4,alpha=1.7)
}
