% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_path_generator.r
\name{a_tilda}
\alias{a_tilda}
\alias{a_tilda_R}
\title{Creates the corresponding value from the paper by Stoev and Taqqu (2004).}
\usage{
a_tilda(N, m, M, alpha, H)

a_tilda_R(N, m, M, alpha, H)
}
\arguments{
\item{N}{a number of points of the lfsm.}

\item{m}{discretization. A number of points between two nearby motion points}

\item{M}{truncation parameter. A number of points at which the integral representing the definition of lfsm is calculated. So, after M points back we consider the rest of the integral to be 0.}

\item{alpha}{self-similarity parameter of alpha stable random motion.}

\item{H}{Hurst parameter}
}
\description{
a_tilda triggers a_tilda_cpp which is written in C++ and essentially
performs the computation of the value.
}
\section{Functions}{
\itemize{
\item \code{a_tilda_R}: Old version of a_tilda function. Written entirely in R. It
was left for comparison, testing and backward compatibility.
}}

\references{
\insertRef{StoevTaqqu04}{rlfsm}
}
